Rename timer files
